Editorial: use "active function object" more often
In #1541, a few checks on the "currently-executing constructor" were
added. The pull request review revealed that this phrase was not
sufficiently precise, so it was replaced with "active function object",
a term defined in the JavaScript specification. However, this was only
done in one of the three places the imprecise term appears. This tidies
up the other two.