Handle error cases during module script tree fetching/running better
As discussed in #1545 and in
https://esdiscuss.org/topic/moduledeclarationinstantiation-behaviour-after-failure,
the spec would potentially call ModuleDeclarationInstantiation on a
module that previously failed to instantiate, with cascading bad
consequences for the rest of the spec's algorithms. The JavaScript spec
expects us to track these failures and avoid calling
ModuleDeclarationInstantiation a second time. This commit adds such
state tracking, partially by moving the ModuleDeclarationInstantiation
calls to fetch time (with any errors stored for later rethrowing during
evaluation time).

Fixes #1545.