Simplify the reentrance guard for custom element definition
Previously, we would allow reentrance as long as it was not with the
same name or constructor as was currently being defined, using the
mechanism of a "set of being-defined names" and "set of being-defined
constructors". This was overly complex, given that there are not really
any use cases for defining a custom element in a getter triggered by the
element definition algorithm. So instead, we move to just using an
overall reentrance-guard flag.

Fixes https://github.com/w3c/webcomponents/issues/545.