Update and correct <meta name="referrer"> parsing
This stems from
https://github.com/w3c/webappsec-referrer-policy/issues/53 and
https://github.com/w3c/webappsec-referrer-policy/pull/66 which remove
the legacy keyword support from the HTTP Referrrer-Policy header. We
instead move that into HTML as a preprocessing step for the <meta
name="referrer"> element, before setting the resulting policy on the
document.

While there, this fixes the <meta name="referrer"> parsing to align with
Chrome's implementation, as seen at
https://chromium.googlesource.com/chromium/src/+/1348a36a3427d39554eab15e8920cec368829355/third_party/WebKit/Source/core/dom/ExecutionContext.cpp#239:

- It includes the additional legacy translation
  "origin-when-crossorigin" => "origin-when-cross-origin".
- It splits the given value on commas and uses the last valid value that
  is found, similar to what is done for the header.
- It makes the search process use case-insensitive matching.