Fix some callers of "create a module script"
This algorithm was not always being called with the correct arguments;
in particular, when creating an inline module script, the cryptographic
nonce, parser state, and credentials mode were not being set correctly.
This fixes that call site and also updates its argument order to be
consistent with other nearby algorithms.