Set "one permitted sandbox navigator" for all sandbox-created popups
This patch ensures that any browsing context created from a sandboxed
document will have its "one permitted sandbox navigator" set.
Previously, this property was not set when navigating from a sandboxed
document whose "sandbox propagates to auxiliary browsing contexts flag"
was not set (e.g. those loaded with 'allow-popups-to-escape-sandbox'),
which meant that sandboxed documents could navigate unrelated windows.
Closing this hole in the spec aligns with the behavior Firefox plans
to ship, and Chrome intends to align with this behavior as well.

Fixes #1218.