Run more of the SharedWorker constructor in parallel
This should make it clearer that all the interprocess communication is
not happening while we run the SharedWorker constructor as that would
be somewhat odd.

It hopefully also helps us in the future as it becomes more obvious we
cannot throw exceptions from code that runs in parallel.

Fixes the remainder of #1819.