Update the parser's element creation for recent DOM changes
This follows up to https://github.com/whatwg/dom/pull/341 (which is in
turn part of https://github.com/w3c/webcomponents/issues/569). Due to
the changes there, it is no longer necessary for HTML to catch and
report the element creation error during parsing; instead, "create an
element" performs that.