Move service worker concept to environment
Address the changes outlined in https://github.com/whatwg/html/pull/1776#issuecomment-250465276 with yet another tweak.

In order to keep the well established semantics around the client's
active worker concept, this patch suggests we store the (active)
service worker instead of the service worker registration in an
environment.

By this change, the main resource and its subresources will be
controlled by the same version of active worker as when the client
associated with the main resource is started to be controlled by a
service worker its containing registration's waiting worker won't
promote until all the clients are unloade (including this reserved
client).

* To reviewer: this message can be squashed away while merging the PR.