Remove canvas element's hit region feature (for now)
As currently defined the feature has a number of objections due to it
altering hit testing without there being a real need to do so. The API
can also be simplified to only work with elements rather than
accommodating arbitrary objects.

Fixes #548. Closes #547, closes #849, and closes #1029, due to hit
regions requiring a fresh PR. #1030 is the follow-up issue for new hit
region work.