Remove tab index's synthetic click event feature
Other than through a feature called “spatial navigation”, currently not
enabled by default in any shipping browser, this behavior is not
implemented and at this point it’s highly likely shipping it would
break compatibility with deployed content as discussed in #1952.

Removing it to reflect reality, but as we expand the features of custom
elements, exposing “activation behavior” somehow should be high on the
list as the effective default is detrimental to keyboard-only users.

Closes #1952.