Meta: Make the "File an issue" link available ASAP
There is no reason to wait with running this script; initialization
is cheap (at least as of [1]). Running it `async` instead of `defer`
makes the link available as soon as possible, so you can file an
issue while the spec is still loading.

Also avoid using the shortlink for this link, since with query
strings it's not short anyway, and the redirect causes extra delay.

[1] https://github.com/whatwg/resources.whatwg.org/pull/23