Editorial: Re-reference RFC 2119 in spec text
All the spec’s citations to RFC 2119 as a reference in the References
section were removed in the 4ac633e refactoring related to the Infra
Standard. But one place remains in the spec that should have already
included a citation to RFC 2119 as a reference in the References
section. So this change adds that where it should be.