Make same-URL-navigation replace the current entry
Fixes #1742.

(There is an enormous amount of cleanup left to be done here. We don’t
currently pass the browsing context correctly between algorithms.
History is poorly defined in terms of global concepts that can come
from anywhere. The URL argument to navigation is actually more
complicated but glossed over in several places. We need a lot of
refactoring here.)