Remove special handling of <rp> in innerText's getter
This was a hack that is only implemented in one engine (Gecko).
For example, an <rp> in <select> is included, which is weird;
if an <rp> has several text nodes, only the first is included.

There is not enough implementor interest in making <ruby> actually
work well in innerText (e.g. in the absense of <rp>); see #1801.