Allow <link rel=apple-touch-icon sizes=… href=…>
This changes the document-conformance requirements for the `link` element
such that it’s conforming for it to have a `sizes` attribute if it has
`rel=apple-touch-icon` (instead of `rel=icon`). But this doesn’t change any
UA requirements; instead it includes a note that (unlike for `rel=icon`),
UAs are not actually required to do anything for `rel=apple-touch-icon`.