Meta/editorial: update how internal slots and methods are defined
This updates how the internal methods and slots defined by this spec,
namely [[Transfer]](), [[Clone]](), [[Detached]], and a few
WindowProxy/Location internal slots, are marked up. To fit better with
other specs in the ecosystem, the brackets are now considered part of
their name, and their Bikeshed metadata has changed to be "methods" and
"attributes" of platform objects.