Clarify valid URL
The URL standard says "A URL is a universal identifier. To disambiguate from a URL string it can also be referred to as a URL record." So it's not correct to refer to this definition of "URL" for the validity requirements of an href attribute, which contains a string, not a record.

It was intended to refer to a "URL string", which is (more or less) the valid input to the URL parser.