Editorial: Drop mention of <hgroup> “masking” subheadings
It’s misleading for the spec to say the purpose of the <hgroup> element is to
“mask” secondary <h1>-<h6> headings from the outline. It’s misleading because
per the outline algorithm, *all* <h1>-<h6> children of an <hgroup> do in fact go
into the *abstract* outline, together, as a unit. No headings are dropped.

And how exactly those secondary headings get shown in *rendered* output of an
outline is an implementation decision to be left up to implementors of tools
that generate rendered views of outlines.

In other words, the spec doesn’t state that any normative requirement that
secondary headings should be “masked” in the rendered view of an outline, nor
does it state any normative requirements at all for rendering of outlines.

So this change drops the mention of <hgroup> being using to “mask” anything.