Define encoder failure actions for toDataURL/toBlob
This is a normative change that defines the proper behaviors of
image file serialization APIs in the event of an encoder error.
For toDataURL, the behavior is to return the string "data:,". For
toBlob, the behavior is to invoke the callback with null. For
OffscreenCanvas.convertToBlob, the behavior is to reject the
promise with and EncodingError exception.