Remove <button type=menu>
This removes the type="menu" option for buttons, which has wide-reaching
consequences for the menu infrastructure.

It was intended that <menu type="context"> (originally <menu
type="popup">) would be used both for context menus, and for popup menus
triggered by buttons. Only the context menu case was ever implemented
(in Gecko and, behind a flag, in Blink). As such, removing support for
the triggered-by-button case allows us to remove a lot of
now-unnecessary abstraction; for example, we can rename the "popup menu
state" to the "context menu state" to match the type="context" content
attribute. And we can relocate all of the processing model for
constructing menus into the context menu processing model section.
Finally, we can clearly state for authors that the <menu> element has
two distinct uses: for customizing the context menu (type="context"),
and as a semantic alternative to <ul> for a list of commands
(type="toolbar").

This fixes #237, although we still need to keep an eye on the entire
context menu feature since its cross-browser support is on the low side.