Define encoder failure actions for toDataURL/toBlob
Define the proper behaviors of image file serialization APIs in the event of an encoder error. For toDataURL(), the behavior is to return the string "data:,". For toBlob(), the behavior is to invoke the callback with null. For OffscreenCanvas's convertToBlob(), the behavior is to reject the promise with a DOMException "EncodingError".

Test tracking issue (since no cross-browser tests can be created): https://github.com/w3c/web-platform-tests/issues/4802.