Refactor the document.domain setter as a standalone algorithm (1/2)
This is in response to W3C/HTML PR https://github.com/w3c/html/issues/769.

The Web Authentication WG's draft currently makes reference to the "Relaxing the
same-origin restriction" of the document.domain attribute setter as a way to let
relying parties use foo.bar.com to generate scoped credentials for bar.com.

However, 1) the attribute setter procedure isn't documented as an algorithm - so
we shouldn't call it like one, and 2) we need to override some of the ambient
state within it, by changing some of the values to be passed as arguments.

We had started some work to inline the procedure as an algorithm within our
document, but consensus is that it'd be better if we could avoid future
divergence by refactoring this part of the HTML spec instead.