Remove the dropzone global attribute.
dropzone failed to get traction among browser implementers. Having it in
the specification is confusing to Web developers who may attempt to use
it, only to discover that it is not supported.

Safari and Chrome implemented a prefixed version, webkitdropzone. The
prefixed version is going away from Chrome [1]. Other browser vendors
have no objections to the attribute getting removed from the spec [2].

[1] https://www.chromestatus.com/feature/5718005866561536
[2] https://github.com/whatwg/html/issues/2331