Allow <link rel=mask-icon color>
This changes the document-conformance requirements for the `link` element
such that it’s conforming for it to have a `color` attribute if it has
`rel=mask-icon` (instead of `rel=icon`). But this doesn’t change any
UA requirements; instead it includes a note that (unlike for `rel=icon`),
UAs are not actually required to do anything for `rel=mask-icon`.