Remove the dropzone global attribute
dropzone failed to get traction among browser implementers. Having it in
the specification is confusing to Web developers who may attempt to use
it, only to discover that it is not supported.

WebKit and Blink implemented a prefixed version, webkitdropzone. The
prefixed version is going away from Blink [1]. WebKit has no objections
to the attribute getting removed from the spec [2].

Fixes #2331. Tests:

* https://github.com/w3c/web-platform-tests/pull/5052
* https://github.com/w3c/web-platform-tests/issues/5053

[1]: https://www.chromestatus.com/feature/5718005866561536
[2]: https://github.com/whatwg/html/issues/2331#issuecomment-280746724