Specify the tokenizer for window.open's features argument
This was specified in CSSOM View but the "noopener" feature did
not use the same tokenizer as the legacy features. Fixes #2474.

Also specify the aliases screenx, screeny, innerwidth, innerheight
for left, top, width, and height, respectively. Part of #2464.
Closes https://github.com/w3c/csswg-drafts/pull/1128.

The tokenizer specified here closely follows Chromium and WebKit.
Difference from Chromium: U+0000 is not a separator.
Difference from WebKit and Chromium: U+000C is a separator.