Editorial: refine link-element section readability
The section of the spec on the link element is long and a bit complicated.
This editorial change does some reorganization to make it more usable/readable
both for implementors and authors. It restructures it into several subsections
for the normative requirements, and reorders content to keep generic link stuff
first, with rel-specific stuff afterward, and grouped by rel.

Before this change the UA requirements for implementors were intermingled among
the document-conformance/informational parts for authors (1, 2, 3 above), and
not called out with headings of any kind—which made the section less
usable/readable both for implementors and authors.

This change doesn’t add or alter any normative requirements; it just moves a few
paragraphs around, adds some headings, and reorders some things.