Cut distracting statement about external resources
The section of the spec that provide details about attributes for the
`link` element contains the following statement:

> Some of the attributes control whether or not the external resource is
> to be applied (as defined below).

This change drops that statement.

That statement has been in the spec for 11 years but isn’t stating a
normative requirement nor serving any useful purpose—instead, it’s
actually just distracting and counter-productive in that despite saying
“as defined below”, the notion of “to be applied” isn’t actually clearly
defined at all in the prose anywhere in the spc below that statement.

Fixes https://github.com/whatwg/html/issues/2503