Change <input type=range> underflow/overflow behavior
This fixes #2427, by only clamping to the minimum/maximum when the
element's dirty value flag is set. This allows attributes to be set
without regard for their order, which is much better than being
order-dependent (and thus does not require magic in the HTML parser to
set multiple attributes at once). It also better matches the informative
definition of the dirty value flag as tracking the relationship between
the element's value and its default value.