Update <link>'s processing model to work well for preload
This allows the preload spec to delegate to HTML's "obtain the resource"
algorithm for elements, instead of creating a copy of it. The
latter approach has been attempted, but got out of date very quickly;
see w3c/preload#91 for more details.

This also updates rel="preload" to account for recent changes to Fetch
that introduced the concept of a "potential destination" for use here.
See https://github.com/whatwg/fetch/pull/547.

Notably, part of this integration is ignoring invalid as="" attributes,
which now includes empty ones. No error event is fired.

Finally, it also allows other specifications to override the delaying of
the load event, as the preload spec does.