Tweak <link>.workerType reflection
9c7da4998f005c1799b04f452cb6c4a618f2d42e removed the idea of using
enumerations in the IDL to case-sensitively reflect properties, but
forgot to switch <link>'s workerType attribute to DOMString. This fixes
that.

Additionally, this introduces the invalid state as the invalid value
default, so that service workers' processing model can ignore invalid
values instead of treating them as classic scripts (similar to
<script>'s type="" attribute).