Meta: Fix bug that loaded xref.json for any click
This change is cleanup for a bug in 3af8da2, which was causing xref.json
to be loaded for any click anywhere in the document other than a
hyperlink. This change corrects the behavior such that xref.json is
loaded only when the click is either on a dfn or descendant of dfn.