Update textarea textContent/defaultValue/value interaction
This fixes #2752, where it was revealed that most browsers do not care
about the textContent of a textarea, but instead about its child text
content. That is, if you use the DOM APIs to insert a child element into
the textarea element, its text does not show up for editing or as part
of the value and defaultValue properties (even though it shows up in
textContent).

This moves to a model where textareas operate entirely on the child text
content. This matches Blink and Gecko, and somewhat matches WebKit.

This also fixes #2750, by using the newly-introduced "child text content
change steps" hook introduced to DOM in
https://github.com/whatwg/dom/pull/466.