Meta: changes to support a revamped developer's edition
This new dev edition is meant to supercede
https://developers.whatwg.org/. It is located at
https://html.spec.whatwg.org/dev/. Closes
https://github.com/benschwarz/developers.whatwg.org/issues/104.

This introduces a developers stylesheet derived from
https://github.com/benschwarz/developers.whatwg.org/tree/97ff943a8f5b8fe38f78e224b4b44b472ccefb57/sass,
including the web fonts. We host the web fonts in this repository,
instead of using Google Fonts, as Google Fonts is blocked in China.
Further help on improving the design of the dev edition going forward is
appreciated; for now we have done a straight port.

It also updates the source document significantly to make the generated
output work better for the dev edition. Special care was given to the
web workers section, and to ensuring that every element definition
references which interface it corresponds to. Further work on properly
marking up the source document is tracked in #1488.

This new version is missing several features from the old one. In rough
order of importance:

* Search
* Working offline
* "About this specification" introduction on the index page
* Nice footer and header navigation UI on each non-index page
* A distinct favicon

We plan to add most of these back over time.