Require text entry cursor at the beginning of controls initially
Previously the spec allowed this to be "up to platform conventions", but
the web depends on it being at the beginning: see
https://bugzilla.mozilla.org/show_bug.cgi?id=1337392.

This also updates the spec to require resetting the cursor to the
beginning after type changes. This is not currently interoperable,
although it is the case in Edge and Gecko.

This was brought up in the context of the larger issue of #2424.