Fix broken step ref in “fire a DND event” algo
A step in the “fire a DND event” algorithm referred back to an action occurring
in the “first step” of the algorithm, but since the time when that reference was
first made, additional steps were apparently added to the beginning of the
algorithm such that what had been the first step is now step 3. So this change
refines the algorithm to instead set and use an internal flag rather than
relying on a hardcoded reference back to a specific step.

Credit to @rniwa for catching it