Fix lost ampersands in Character Reference State's Anything Else case.
The Anything Else case of the Character Reference State discarded the
ampersand that had been consumed to enter the state. This led to e.g. in
the Data State, "&*" emitting only a "*" char token, when you'd expect
both a "&" and "*" char token.

This fix appends the & to the attribute value if it was consumed while
consuming an attribute value, or otherwise emits it as a character token.