Regression fix: Lost ampersands in Character Reference State's Anything Else case
The Anything Else case of the Character Reference State discarded the
ampersand that had been consumed to enter the state. This led to e.g. in
the Data State, "&*" emitting only a "*" char token, when you'd expect
both a "&" and "*" char token.

This regressed in ee198945e388bd3298b73fb842fa0175d5b6a0f9.