Meta: move deploy process to Travis
This builds and validates the spec entirely on Travis, and then deploys
it to the WHATWG servers. Previously Travis was only running build and
validation steps, but then throwing away the results, letting custom
deploy architecture handle the actual deploy. This replaces the custom
deploy architecture with something in source control and
web-host-agnostic.

This deploy is Docker-based, for two main reasons:

* It is not easy to install FreePascal 3.x onto Travis CI's default
  Ubuntu configuration, which is fairly old (even with recent Trusty
  updates). It is simpler to create a Docker container with a recent OS
  and install FreePascal 3.x there.
* Docker has a good mechanism for caching previous results and not doing
  unnecessary work until the cache is invalidated. This is important as
  it allows us to avoid reinstalling prerequisite packages and
  recompiling Wattsi on every build. Travis CI has some caching support,
  but it is not as full-featured as Docker's. The intermediate
  containers created for this caching are stored on Docker Hub; see
  https://hub.docker.com/r/whatwg/html-deploy.

While here, we update the Travis configuration to no longer runs Travis
on both pull requests and branches, but instead only on pull requests.