Bring outdated parts of the FAQ up to date
This change touches older parts of the FAQ that have become outdated, making
adjustments to the wording of some parts to bring them into alignment with the
language in the current spec, and — for a couple of other parts that are no
longer as relevant — moving them into a "Legacy questions" section. Some general
copy-editing refinements are also included in the change.