Clarify template content model and syntax
This change adds a link in the `template` *Content model* header
directly to the related example in the prose of the `template` section
that helps make clear why the content model is `Nothing`.

This change also adds the `template` element as a new (sixth) element
type in the HTML Syntax: Writing HTML documents: Elements section.

Fixes #3021.