Throw when resizing OffscreenCanvas placeholder
Attempting to set the width or height attribute of a canvas element
after it has transferred control to an OffscreenCanvas should throw an
exception. Before this change, the behavior was to silently ignore
changes to these attributes, which was not clear because that case was
not specifically mentioned in the spec.