Add text-decoration-skip default value to UA stylesheet
Following the discussions in [1] we propose to add the value "objects
ink" to the body, html elements for the text-decoration-skip property. I
interpret the thread to have reached consensus on the aspect of making
ink skipping a default behavior by means of a UA stylesheet. We believe
it brings better typography to the web.

[1] https://github.com/w3c/csswg-drafts/issues/727#issuecomment-268945879