Refine the inputmode attribute, and expose it on contenteditable
Move the inputmode attribute definition into the editing section
so that we can declare it in the ContentEditable IDL, and allow it as a
global attribute. Closes #1897.

Remove the prose concept from inputmode (ie. latin, kana) since it
hasn't been implemented and is not possible to implement in current
virtual keyboard implementations. Closes #3077, by resolving to do this
subsetting instead of full removal. Similarly closes #1626 by
establishing inputmode="numeric" as the solution.

Tests: https://github.com/w3c/web-platform-tests/pull/8690