Modify OffscreenCanvas.ConvertToBlob spec
OffscreenCanvas.ConvertToBlob is modifed so as to be consistent with
the spec of canvas.toBlob. In particular, when the canvas element's
bitmap has no pixels, the function should not throw an exception;
instead, it should let the result blob be null. It's the similar for
the case when the encoding is not successful.