Simplify getContext() and make it throw less
In practice getContext() only has two arguments. And in practice some
handling of the second argument is shared across the various rendering
context types. This aligns the standard with that practice.

This also obviates the need for dedicated coercion algorithms for "2d"
and "bitmaprender".

This change also fixes various minor nits found along the way, such as
the inconsistent spelling of contextId (was sometimes contextType).

Tests: https://github.com/w3c/web-platform-tests/pull/9194 and
https://github.com/w3c/web-platform-tests/pull/9205.

Fixes #595.